+++ changelog.md
@@ -8,7 +8,39 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Added
+
+- Usability: QuickSearch - Select first on enter
+  [discussion #328](https://github.com/e-netze/webgis-community/discussions/328)
+  [docs](https://docs.webgiscloud.com/de/webgis/apps/viewer/customjs/usability.html#schnellsuche)
+
 ### Fixed
+
+## 7.25.3601
+
+### Added
+
+- WMS Service: SLD_VERSION Support
+  [discussion #279](https://github.com/e-netze/webgis-community/discussions/324)
+
+- Usability: Show a note in query table, if the corresponding layer is not visible in the map
+  [discussion #334](https://github.com/e-netze/webgis-community/discussions/334)
+  can be switched on/off in ``custom.js``
+
+  ``` javascript
+    // default: false for all API apps
+    // but set 'true' in custom-recomandations.js for the Viewer
+    webgis.usability.showQueryLayerNotVisbleNotification = false;
+  ```
+
+- Usability: Show Section length/area with measure tools
+  [discussion #325](https://github.com/e-netze/webgis-community/discussions/325)
+
+- MapMarkup: Added tool for measure polygon area/circumference
+
+### Fixed
+
+- Bug: MapMarkup: download GPX/Shape
+  [discussion #335](https://github.com/e-netze/webgis-community/discussions/335)
 
 ## 7.25.3401